Вывод кнопок в зависимости от роли

+++ resources/views/news/index.blade.php
@@ -4,11 +4,12 @@
     <div class="container">
         <h1>
             Новости
-            <a class="btn btn-primary pull-right"  href="{{ route('news.create') }}">Добавить</a>
+            @if($roles->contains('moderator'))
+                <a class="btn btn-primary pull-right"  href="{{ route('news.create') }}">Добавить</a>
+            @endif
         </h1>
         <hr>
         @include('news.info')
-
         <div class="row">
             <div class="panel panel-default">
                 <div class="panel-body">
@@ -16,11 +17,17 @@
                         <div class="row">
                             <div class="col-md-8"><a href="{{ route('news.show', $item->slug) }}">{{ $item->title }}</a></div>
                             <div class="col-md-2">{{ $item->created_at }}</div>
-                            <div class="col-md-1"><a class="btn btn-warning" href="{{ route('news.edit', $item->slug) }}">Edit</a></div>
                             <div class="col-md-1">
-                                {!! Form::open(['route' => ['news.destroy', $item->slug], 'method' => 'DELETE']) !!}
-                                <button class="btn btn-danger">Del</button>
-                                {!! Form::close() !!}
+                                @if($roles->contains('moderator'))
+                                    <a class="btn btn-warning" href="{{ route('news.edit', $item->slug) }}">Edit</a>
+                                @endif
+                            </div>
+                            <div class="col-md-1">
+                                @if($roles->contains('admin'))
+                                    {!! Form::open(['route' => ['news.destroy', $item->slug], 'method' => 'DELETE']) !!}
+                                    <button class="btn btn-danger">Del</button>
+                                    {!! Form::close() !!}
+                                @endif
                             </div>
                         </div>
                         <hr>
